Backend/requirements.txt
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
langdetect==1.0.9
openai==1.35.5
python-dotenv==1.0.1
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.1
